% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_screening.R
\name{add_to_screening}
\alias{add_to_screening}
\title{Adds new articles to ongoing screening process}
\usage{
add_to_screening(old_article_df, new_article_df, team_df, seed = 1)
}
\arguments{
\item{old_article_df}{df that contains all work-in-progress data, retrieved by get_screening_sheets()}

\item{new_article_df}{a df that contains new articles, and has only one identifier}

\item{team_df}{data frame describing the team that contains name<chr>, and screening_effort<int> variables to be used when assigning the new articles}

\item{seed}{random seed to use to make the assignment reproducible}
}
\value{
A data frame containing nested data frames for each revier with updated data
}
\description{
Add new literature to the existing and work-in-progress files and return an updated df. It preserves all previous screening effort, so no manual coding is lost.
}
\examples{
add_to_screening(old_articles, psycinfo_df, team, 1)
}
