% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_articles.R
\name{assign_articles}
\alias{assign_articles}
\title{Assign (or re-assign) articles to several team members}
\usage{
assign_articles(df = NULL, team_df = NULL, effort_col = NULL, reviewers = 2, reassign = FALSE, seed = 1)
}
\arguments{
\item{df}{data frame of articles, should have at least title and abstract columns. If the function is used for re-assigning, then there should be a decision, name, and position column as well.}

\item{team_df}{a data.frame of team members with name<chr>, and one or more effort <dbl> columns}

\item{effort_col}{variable name <chr> in the team file that specifies the effort of members}

\item{reviewers}{number of team members <num> for each article}

\item{reassign}{is it a reassigning of studies <lgl>?}

\item{seed}{a random seed <num> for reproducibility}
}
\value{
A data frame that contains article data assigned to multiple team members for manual data entry.
}
\description{
Randomly assign articles to any number of team members for screening. The function will create a nesting variable (reviewer) that can be used to nest and save data.
}
\details{
If the manual coding has already started, and some of the articles have been already processed, reassign = TRUE will preserve the work that has been done, and only assign unprocessed articles.
}
\examples{
library(magrittr)
library(tibble)

team_df <- tibble(name = c("Tom", "Jerry", "Rose"),
                  screening_effort = c(.6, .3, .1))

workaholism_pubmed \%>\%
  # First, keep only one identifier for each article
  make_id(identifier = c("doi", "pmid")) \%>\%
  assign_articles(team_df = team_df,
                  reviewers = 2,
                  effort_col = "screening_effort")
TODO: Refactor to remove redundancy (assigning is there twice)
TODO: tidyevalize the effort parameter
}
