% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_screener_irr.R
\name{calculate_screener_irr}
\alias{calculate_screener_irr}
\title{Calculate inter-rater agreement statistics beween team members}
\usage{
calculate_screener_irr(articles, kr_method = "nominal")
}
\arguments{
\item{articles}{a data frame that contains the screened articles, and having a name<chr> column}

\item{kr_method}{The scale of the decision variable. Can be nominal (default), ordinal, interval, or ratio}
}
\value{
A data frame containing agreement summaries, and the Krippendorff's alpha statistic for all pairs of screeners. invalid_decision<int> is the total number of decisions that are not 1 or 0.
}
\description{
Calculate agreement summaries and Krippendorff's alpha for all reviewer pairs. Krippendorff's alpha can be calculated for nominal, ordinal, interval, and ratio scale variables.
}
\examples{
calculate_screener_irr(articles)
}
