% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path_structure.R
\name{create_path_structure}
\alias{create_path_structure}
\title{Create a full path structure}
\source{
Cumulative paste function is based on https://stackoverflow.com/questions/24862046/cumulatively-paste-concatenate-values-grouped-by-another-variable created by https://stackoverflow.com/users/2414948/alexis-laz
}
\usage{
create_path_structure(folders = NULL)
}
\arguments{
\item{folders}{a vector of paths <chr> that does not have to contain all folder levels separately}
}
\value{
a vctor <chr> of paths that fully define a folder structure
}
\description{
Create a full hierarchical path structure based on a vector of folder names with incomplete hierarchical definition
}
\examples{
# Creating default folder structure
create_path_structure(c("research/meta/extraction","research/meta/screening"))
}
