% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_duplicate_id.R
\name{flag_duplicate_id}
\alias{flag_duplicate_id}
\title{Flag duplicates by identifiers}
\usage{
flag_duplicate_id(df, keys = NULL)
}
\arguments{
\item{df}{a data frame with potential duplicates}

\item{keys}{a character vector of identifier variables in the data frame}
}
\value{
The original data frame is returned with a new column "duplicate_by_id", that can be 0 or 1
}
\description{
Flags articles in a dataframe that are duplicated and has the same identifier
}
\examples{
flag_duplicate_id(df, c("doi","pmid","psyid","eid"))
}
\seealso{
\code{\link{flag_duplicates_by_title}} for flagging by title
}
