% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_duplicate_title.R
\name{flag_duplicate_title}
\alias{flag_duplicate_title}
\title{Flag duplicates by title}
\usage{
flag_duplicate_title(df, title = "title", max_distance = 5L)
}
\arguments{
\item{df}{a data frame with potential duplicates}

\item{title}{the <chr> column in df  that has the title of studies}

\item{max_distance}{the maximum difference of titles to be flagged as duplicate}
}
\value{
The original data frame augmented with "duplicate_by_title" column, that can be 0 or 1
}
\description{
Flags articles in a dataframe that are duplicated and has the same identifier
}
\details{
The function uses Optimal String Alignment distance to find the difference between strings, using the stringdist::stringdist() function (for details, see \code{\link[stringdist]{stringdist-metrics}}). Note that this function can also be used to find duplicates based on the abstract or any other text field. It can be a computationally heavy task for more and longer strings.
}
\examples{
library(dplyr)
# Show all articles with duplicated title
merge_sources(workaholism_pubmed, workaholism_psychinfo) \%>\%
 make_id(c("psyid", "pmid", "doi", "eid", "sid")) \%>\%
 flag_duplicate_title(title = "title") \%>\%
 filter(duplicate_by_title == 1)
}
