% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_duplicates_by_title.R
\name{flag_duplicates_by_title}
\alias{flag_duplicates_by_title}
\title{Flag duplicates by title}
\usage{
flag_duplicates_by_title(df, title = NULL, max_distance = 4L, ...)
}
\arguments{
\item{df}{a data frame with potential duplicates}

\item{title}{the <chr> column in df  that has the title of studies}

\item{max_distance}{the maximum difference of titles to be flagged as duplicate}

\item{...}{parameters to be passed to stringdist::stringdistmatrix()}
}
\value{
The original data frame augmented with "duplicate_by_title" column, that can be 0 or 1
}
\description{
Flags articles in a dataframe that are duplicated and has the same identifier
}
\details{
Note that this function can also be used to find duplicates based on the abstract or any other unique text field. However, this might prove to be slow due to the string distance calculations.
}
\examples{
flag_duplicates_by_title(df, "title", max_distance = 3)
}
\seealso{
\code{\link{flag_duplicates_by_ids}} for flagging by ids
}
