% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cellwise_conversion_fails.R
\name{get_cellwise_conversion_fails}
\alias{get_cellwise_conversion_fails}
\title{Cell-wise warnings for conversion failiures}
\usage{
get_cellwise_conversion_fails(df, fun)
}
\arguments{
\item{df}{data frame}

\item{fun}{unquoted conversion function name without ()}
}
\value{
A data frame, where cells contain either NA (no problem with conversion), or the warning message as string
}
\description{
Helper function to get conversion warnings separately for each cell
}
\examples{
df <- tibble::tibble( a = as.character(1:5),
                      b = as.double(1:5),
                      c = as.character(c(1:2, letters[1:3])),
                      d = as.character(c(letters[4:5], c(3:5))))
get_cellwise_conversion_fails(df, rlang::as_integer)
}
