% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversion_fails.R
\name{get_conversion_fails}
\alias{get_conversion_fails}
\title{Checks for coversion errors in a dataframe}
\usage{
get_conversion_fails(df, columns = NULL, convert_to = c("integer","double","logical"))
}
\arguments{
\item{df}{data frame}

\item{columns}{a vector of column names <chr> to be converted}

\item{convert_to}{output variable type can be "integer", "logical", "double" (character and factor conversion does not produce errors)}
}
\value{
A tibble that shows all conversion errors for a type
}
\description{
Checks if a specific conversion produces any errors in a df.
}
\examples{

df <-
  tibble::tibble(file = c("file1","file1", "file2", "file2", "file2"),
                 x = c(1,2,"c", 4, "d"),
                 y = c(1, 3, NA, 23, 3.0, 0)) \%>\%
  dplyr::group_by(file) \%>\%
  tidyr::nest()

get_conversion_fails(df, columns = c("x", "y"), convert_to = "integer")
TODO: change example
}
