% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_folder.R
\name{get_from_folder}
\alias{get_from_folder}
\title{Get csv-s from folder}
\usage{
get_from_folder(local_path = NULL, all_char = TRUE)
}
\arguments{
\item{local_path}{A google drive path where the sheets are stored}

\item{all_char}{Read all varaibles as characters? (no data loss)}
}
\value{
The content of all spreadsheets in a local folder in one nested tibble
}
\description{
Reads all csv-s from a local folder into one nested data frame
}
\examples{
\dontrun{
get_from_folder("Research/meta-analysis12/Extraction1/")
}
}
