% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gdrive.R
\name{init_gdrive}
\alias{init_gdrive}
\title{Create folder structure on goodle drive}
\usage{
init_gdrive(gdrive_path = NULL, folders = c("literature_search/", "screening/", "screening_consensus/", "fulltext/", "extraction/", "extraction_consensus/"))
}
\arguments{
\item{gdrive_path}{a full (new) gdrive path <chr>, preferably with / at the end}

\item{folders}{a vector <chr> of folder names to create in path}
}
\value{
no output, this function exerts a side-effect
}
\description{
Create a folder structure (default or custom) on google drive for a meta-analysis project
}
\examples{
# Creating default folder structure
init_gdrive("research/meta-analysis/")
# Creating custom folder structure
init_gdrive("research/meta-analysis/", c("screening", "extract"))
}
