% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_meta.R
\name{init_meta}
\alias{init_meta}
\title{Create a folder structure and a team file on goodle drive}
\usage{
init_meta(gdrive_path = NULL)
}
\arguments{
\item{gdrive_path}{a full (new) gdrive path <chr>, preferably with / at the end}
}
\value{
no output, this function exerts a side-effects
}
\description{
Create a folder structure on google drive for a meta-analysis project and a template team file in it that opens in the browser
}
\examples{
# Creating default folder structure and team file
init_meta("research/meta")
# TODO: A bug makes the function to initialize the gdrive_path directory twice, therefor it breaks halfway
}
