% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sources.R
\name{merge_sources}
\alias{merge_sources}
\title{Merge data sources}
\usage{
merge_sources(..., .renames = NULL, .tidy_names = TRUE, .all_char = TRUE)
}
\arguments{
\item{...}{names of data dataframes to be merged}

\item{.renames}{a named vector that contatins all renames for all datafiles}

\item{.tidy_names}{should the names become tidy?}

\item{.all_char}{convert all variables to character?}
}
\value{
a dataframe that contains all data sources from all files
}
\description{
Merge all sources into a single data frame with tidy names. A named vector can be provided for renaming variables.
}
\details{
For renames, c("new_name" = "old_name") format should be used.
}
\examples{
merge_sources(workaholism_psychinfo, workaholism_pubmed, workaholism_scopus,
                        .renames = c(journal = "publication"))
}
