% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamanager.R
\docType{package}
\name{metamanager-package}
\alias{metamanager-package}
\alias{metamanager}
\title{Functions for efficient meta-analysis project management}
\description{
Helps to menage files and workflow in meta-analysis projects
}
\author{
Tamas NAGY <nagytamas.hungary@gmail.com>
}
