% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_conversion_fails.R
\name{report_conversion_fails}
\alias{report_conversion_fails}
\title{Checks for all coversion errors in nested dataframes}
\usage{
report_conversion_fails(df, columns, convert_to)
}
\arguments{
\item{df}{data frame of nested data frames}

\item{columns}{a list of vectors that contain column names <chr> to be converted}

\item{convert_to}{a list of outcput variable types for each vector of columns. Can be "integer", "logical", "double" (character and factor conversion does not produce errors)}
}
\value{
A tibble that shows all conversion fails for all type conversions for each file
}
\description{
Checks if converting the variables to the appropriate conversion produces any errors in the nested dfs
}
\examples{

df <-
  tibble::tibble(file = c("file1","file1", "file2", "file2", "file2"),
                 x = c(1,2,"c", 4, "d"),
                 y = c(1, 3, NA, 23, 3.0, 0)) \%>\%
  dplyr::group_by(file) \%>\%
  tidyr::nest()

report_conversion_fails(df, columns = c("x", "y"), convert_to = "integer")
TODO:
}
