% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_kripp.R
\name{tidy_kripp}
\alias{tidy_kripp}
\title{Tidy function for Krippendorff's alpha}
\usage{
tidy_kripp(kripp)
}
\arguments{
\item{kripp}{a list object produced by irr::kripp.alpha()}
}
\value{
A tibble containg the values of the Krippendorf's alpha output
}
\description{
Creates a tidy data frame from Krippendorff's alpha list output
}
\examples{
tidy_kripp(screener_12_kripp)
}
