% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_gdrive.R
\name{upload_to_gdrive}
\alias{upload_to_gdrive}
\title{Upload all .csv-s as google sheets from a local folder to a google drive folder}
\usage{
upload_to_gdrive(local_path = NULL, gdrive_path = NULL, overwrite = FALSE)
}
\arguments{
\item{local_path}{a full path <chr> that contains .csv-s to upload to gdrive}

\item{gdrive_path}{a full gdrive path <chr>, preferably with / at the end}

\item{overwrite}{overwrite files <lgl>? (default = FALSE)}
}
\value{
no output, this function exerts a side-effect
}
\description{
The function reads all files from a google drive folder and copies the files to a gdrive folder
}
\examples{
\dontrun{
upload_to_gdrive("/temporary_files/",
                 "temp_files/",
                 overwrite = TRUE)
}
}
