% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workaholism_psychinfo.R
\docType{data}
\name{workaholism_psychinfo}
\alias{workaholism_psychinfo}
\title{Workaholism articles from PsychINFO}
\format{\describe{
  \item{psyid}{PsycINFO identifier}
  \item{pmid}{Pubmed id}
  \item{doi}{Digital object identifier}
  \item{title}{Title of the article}
  \item{abstract}{Abstract of the article}
  \item{year}{Year of publication}
  \item{month}{Month of publication}
  \item{day}{Day of publication}
  \item{journal}{Name of the journal}
  \item{authors}{Name of all authors}
  }}
\usage{
workaholism_psychinfo
}
\description{
Article metadata returned for workaholism keywords by PsychINFO.
Search expression to obtain data: "work addict*" OR "work depend*" OR workaholism
The search was conducted on 2018/04/02
}
\keyword{datasets}
