% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workaholism_pubmed.R
\docType{data}
\name{workaholism_pubmed}
\alias{workaholism_pubmed}
\title{Workaholism articles from PubMed}
\format{\describe{
  \item{pmid}{Pubmed id}
  \item{doi}{Digital object identifier}
  \item{title}{Title of the article}
  \item{abstract}{Abstract of the article}
  \item{year}{Year of publication}
  \item{journal}{Name of the journal}
  \item{authors}{Name of all authors}
  }}
\usage{
workaholism_pubmed
}
\description{
Article metadata returned for workaholism keywords by PubMed.
Search expression to obtain data: ((("work addict*"[Title/Abstract] OR "work depend*"[Title/Abstract] OR workaholism[Title/Abstract]) AND English[lang] AND (Journal Article[ptyp] OR pubmed books[filter])))
The search was conducted on 2018/04/02
}
\keyword{datasets}
