% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns.R
\name{add_columns}
\alias{add_columns}
\title{Add named empty columns to a dataframe}
\usage{
add_columns(df, columns)
}
\arguments{
\item{df}{data frame of articles}

\item{columns}{a vector of column names <str> to be added}
}
\value{
A data frame that contains the article info with empty columns for manual data entry
}
\description{
Add columns to a dataframe that can be used to manual coding
}
\examples{
add_columns(articles, c("decision","reason"))
}
