% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_articles.R
\name{assign_articles}
\alias{assign_articles}
\title{Assign articles for screening to raters reproducibly}
\usage{
assign_articles(df, team_df, effort_col, reviewers = 2, seed = 1)
}
\arguments{
\item{df}{data frame of articles, should have at least title<str> and abstract<str> columns}

\item{team_df}{a data.frame of team members with name<chr>, and one or more effort <dbl> columns}

\item{effort_col}{variable name <chr> in the team file that specifies the effort of members}

\item{reviewers}{Number of team members <num> for each article}

\item{seed}{a random seed <num> for reproducibility}
}
\value{
A data frame that contains the article info with variables to use for manual screening, and the assigned reviewers
}
\description{
Randomly assign articles to two different team members for screening. The function will create files for each team member. The files will contain article data with the first two rows being the decision and reason columns, that can be filled manually during the screening.
}
\examples{
# Assign articles for screening to 2 reviewers
assign_articles(iris, team_df, "screening_effort", reviewer = 2, seed = 1)
}
