% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_categories.R
\name{correct_categories}
\alias{correct_categories}
\title{Match strings with a pre-defined set of strings}
\usage{
correct_categories(to_be_corrected, correct_terms)
}
\arguments{
\item{to_be_corrected}{vector containing the strings to be corrected}

\item{correct_terms}{string vector containing the correct terms}
}
\value{
A corrected string vector that can only contain the correct terms
}
\description{
Correct strings to pre-defined strings. This is a wrapper for stringdist approimate string matching, where certain parameters are preset, and can be used easily in a tidyverse pipe. Using cosine matching to disregard word order.
}
\examples{
articles \%>\% mutate(correct_reason = correct_categories(reason, read_lines("vg-meta-reasons.txt")))
}
