% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_gdrive.R
\name{get_from_gdrive}
\alias{get_from_gdrive}
\title{Get sheets from google drive}
\usage{
get_from_gdrive(gdrive_path = NULL)
}
\arguments{
\item{gdrive_path}{A google drive path where the sheets are stored}
}
\value{
The content of all spreadsheets in the google drive folder in one nested tibble
}
\description{
Reads all google sheets from a google drive path into one nested data frame
Authentication in handled by the googledrive package
}
\examples{
get_from_gdrive("Research/meta-analysis12/Extraction1/")
}
